Move nix objects out of pycore submodule
HDF5 object wrappers (h5group and h5dataset) submodule is now called
HDF5.
