+++ nixio/dimensions.py
@@ -10,17 +10,14 @@ from numbers import Number
 
 import numpy as np
 
-from ..value import DataType
-from ..dimension_type import DimensionType
+from .value import DataType
+from .dimension_type import DimensionType
 from . import util
 
 
 class Dimension(object):
 
     def __init__(self, h5group, index):
-        from nixio.pycore.h5group import H5Group
-        if not isinstance(h5group, H5Group):
-            raise Exception
         self._h5group = h5group
         self.dim_index = index
