+++ nixio/entity.py
