+++ nixio/entity_with_metadata.py
