+++ nixio/group.py
@@ -7,12 +7,12 @@
 # LICENSE file in the root of the Project.
 from __future__ import (absolute_import, division, print_function)
 
-from .pycore.entity_with_sources import EntityWithSources
-from .pycore.data_array import DataArray
-from .pycore.tag import Tag
-from .pycore.multi_tag import MultiTag
+from .entity_with_sources import EntityWithSources
+from .data_array import DataArray
+from .tag import Tag
+from .multi_tag import MultiTag
 
-from .pycore import util
+from . import util
 from .util.proxy_list import RefProxyList
