--- nixio/pycore/__init__.py
@@ -1,7 +1,0 @@
-from .property import Property
-from .section import Section
-from .data_array import DataArray
-from .dimensions import RangeDimension, SetDimension, SampledDimension
-from .feature import Feature
-from .tag import Tag
-from .multi_tag import MultiTag
