--- nixio/pycore/util/__init__.py
@@ -1,3 +1,0 @@
-from .util import *
-from . import names
-from . import units
