+++ nixio/util/__init__.py
@@ -1,0 +1,3 @@
+from .util import *
+from . import names
+from . import units
